refactor: remove redundant code for web view component
