--- src/containers/Browser.js
@@ -1,10 +1,0 @@
-import React, { Component } from 'react';
-import WebPage from '../components/WebPage';
-
-export default class Browser extends Component {
-  render() {
-    return (
-      <WebPage {...this.props} />
-    );
-  }
-}
